Move class to correct package
